Add shared chunk.js file
